requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
beautifulsoup4==4.12.3
lxml==5.3.0
aiohttp==3.11.10
chromadb==0.5.23
openai==1.57.4
python-dotenv==1.0.1
requests==2.32.3
numpy==2.2.1